reorganize project folders
